Crate bip_dht

Implementation of the Bittorrent Mainline Distributed Hash Table.

Modules

message

Structs

DhtBuilder
Stores information for initializing a DHT.

MainlineDht
Maintains a Distributed Hash (Routing) Table.

Enums

DhtEvent
Event that occured within the DHT which clients may be interested in.

Router
Enumerates different routers that can be used to bootstrap a dht.

ShutdownCause
Event that occured within the DHT which caused it to shutdown.

Traits

Handshaker
Trait for peer discovery services to forward peer contact information and metadata.

Type Definitions

InfoHash
Bittorrent InfoHash.

PeerId
Bittorrent PeerId.