Struct bip_dht::MainlineDht

pub struct MainlineDht { /* fields omitted */ }

Maintains a Distributed Hash (Routing) Table.

Methods

impl MainlineDht

fn search(&self, hash: InfoHash, announce: bool)

Perform a search for the given InfoHash with an optional announce on the closest nodes.

Announcing will place your contact information in the DHT so others performing lookups for the InfoHash will be able to find your contact information and initiate a handshake.

If the initial bootstrap has not finished, the search will be queued and executed once the bootstrap has completed.

fn events(&self) -> Receiver<DhtEvent>

An event Receiver which will receive events occuring within the DHT.

It is important to at least monitor the DHT for shutdown events as any calls after that event occurs will not be processed but no indication will be given.